+++ packages/amplify-graphql-generation-transformer/src/index.ts
@@ -1,0 +1,1 @@
+export { GenerationTransformer } from './grapqhl-generation-transformer';