+++ packages/amplify-graphql-transformer-test-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.6.1...@aws-amplify/graphql-transformer-test-utils@1.0.1) (2024-09-06)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils
+
 ## [0.6.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.6.0...@aws-amplify/graphql-transformer-test-utils@0.6.1) (2024-08-28)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils