+++ packages/graphql-transformer-common/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.1](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-common@5.0.0...graphql-transformer-common@5.0.1) (2024-09-06)
+
+**Note:** Version bump only for package graphql-transformer-common
+
 # [5.0.0](https://github.com/aws-amplify/amplify-category-api/compare/graphql-transformer-common@4.31.1...graphql-transformer-common@5.0.0) (2024-08-20)
 
 - feat!: bump version for v2 transformer packages ([7dac35c](https://github.com/aws-amplify/amplify-category-api/commit/7dac35cceb971b256b5688b0745ef82afc78b641))